# integration-tests/grouped-basepath/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BasePath Grouped App Test API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/users/{id}": {
      "get": {
        "summary": "Auto-generated GET /v1/users/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Default fallback response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "userId"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "BasePath Grouped App"
        ]
      }
    }
  }
}